Extract GitHubUserAdapter
